+++ package.json
@@ -12,6 +12,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "bootstrap": "^5.3.7",
     "vue": "^3.5.18",
     "vue-router": "^4.5.1"
   },
