+++ src/components/SideBar.vue
@@ -23,16 +23,25 @@
         <i class="fas fa-user-plus"></i>
         <span>Insertion Joueur</span>
       </router-link>
+      <hr>
+        <router-link 
+          to="/liste-seance" 
+          class="nav-item"
+          active-class="active"
+        >
+          <i class="fas fa-calendar-check"></i>
+          <span>Liste des Séances</span>
+        </router-link>
 
-      <router-link 
-        to="/seance" 
-        class="nav-item"
-        active-class="active"
-      >
-        <i class="fas fa-calendar-alt"></i>
-        <span>Séance</span>
-      </router-link>
-
+        <router-link 
+          to="/insertion-seance" 
+          class="nav-item"
+          active-class="active"
+        >
+          <i class="fas fa-plus-circle"></i>
+          <span>Insertion Séance</span>
+        </router-link>
+<hr>
       <router-link 
         to="/presence" 
         class="nav-item"
@@ -41,7 +50,7 @@
         <i class="fas fa-clipboard-check"></i>
         <span>Présence</span>
       </router-link>
-
+      <hr>
       <router-link 
         to="/match" 
         class="nav-item"
@@ -50,7 +59,7 @@
         <i class="fas fa-futbol"></i>
         <span>Match</span>
       </router-link>
-
+      <hr>
       <router-link 
         to="/composition" 
         class="nav-item"
@@ -59,6 +68,7 @@
         <i class="fas fa-chess"></i>
         <span>Composition</span>
       </router-link>
+      <hr>
     </nav>
   </aside>
 </template>
