+++ src/router/index.js
@@ -2,7 +2,8 @@
 import { createRouter, createWebHistory } from 'vue-router'
 import ListeJoueurs from '@/views/ListeJoueur.vue'
 import InsertionJoueur from '@/views/InsertionJoueur.vue'
-import SeanceEntrenementMatch from '@/views/SeanceEntrenementMatch.vue'
+import ListeSeance from '@/views/ListeSeance.vue'
+import InsertionSeance from '@/views/InsertionSeance.vue'
 import ListePresence from '@/views/ListePresence.vue'
 import ListeMatch from '@/views/ListeMatch.vue'
 import ListeComposition from '@/views/ListeComposition.vue'
@@ -19,9 +20,14 @@ const routes = [
     component: InsertionJoueur
   },
   {
-    path: '/seance',
-    name: 'seanceEntrenementMatch',
-    component: SeanceEntrenementMatch
+    path: '/liste-seance',
+    name: 'listeSeance',
+    component: ListeSeance
+  },
+  {
+    path: '/insertion-seance',
+    name: 'insertionSeance',
+    component: InsertionSeance
   },
   {
     path:'/presence',
